+++ tests/e2e/strategies/scripted_emitter.py
@@ -52,7 +52,7 @@ from itrader.core.sizing import (
     TradingDirection,
 )
 from itrader.strategy_handler.base import Strategy
-from itrader.strategy_handler.config import BaseStrategyConfig
+from itrader.config import BaseStrategyConfig
 
 
 class ScriptedEmitter(Strategy):
