+++ tests/integration/test_universe_spans.py
@@ -28,7 +28,7 @@ import pandas as pd
 
 from itrader.core.sizing import FixedQuantity, SignalIntent, TradingDirection
 from itrader.strategy_handler.base import Strategy
-from itrader.strategy_handler.config import BaseStrategyConfig
+from itrader.config import BaseStrategyConfig
 from itrader.trading_system.backtest_trading_system import TradingSystem
